Retrieve all channels from a given playlist
The TvXmlParser class from Google is a savior in the sense where
we possibly could get all channels without any problem.

The problem is, it only creates channels for the ones that are
defined in the xml. This causes a problem where other channels
in a given playlist can't be present since they might not have
their own given Channel.

Let's parse first the channels with entries in the EPG and then
add the missing channels that are still defined in the M3U
playlist.

AceChannelUtil is the helper class that will ensure everything
related to the playlists.