AceTvInputSetupFragment: Adjust the input id
Upon syncing channels, ChannelSetupFragment needs to have
a correct Input Id in order to retrieve the state of the EPG sync.

Because we were sending null, it couldn't get updated
accordingly.

Fix this by storing the input id into a variable and then use it
appropriately for the EPG sync/getter method.