AcePlayer/AceInputTvService: Show output information
An example shown in the Android TV sample app is that you
can have the information concerning the video (aspect ratio,
resolution, etc) output, the audio output (stereo) and the
subtitles.

With Ace having many similar channels at different resolutions,
it might be better to distinguish each of them since the names
are replaces by logos.

Since it is seen as a TvTrackInfo, get all the information needed
based on the formats decoded by SimpleExoPlayer. Then, force
the default formats into the Session.

Also implement ExoPlayer.EventListener inside the Session so
we can react more realistically to it (For example calling
notifyVideoAvailable() directly when the play is ready instead
of calling it directly after play())