AceChannelUtil: Remove video format when creating a Channel
The only reason the video format was present on the channel
was because it was thought that it could have brought the
information concerning a stream directly to the screen.

However, since this is controlled by TvTrackInfo and that some
channels had wrong formats based on their origin country,
it might be better to remove it altogether.