Create an appropriate channel original network id
Basing ourselves using the EPG's id for a channel might be
good but it is creating channel duplicates since TvContractUtils
uses a SparseArray. The array cannot find the unique key and
thus recreates a new channel.

Fix this by using the channel display name's hash instead of it's
EPG's id hash to create the original network id.

At the same time, let's follow the guidelines by removing
deleteChannelsFromContentProvider() since it was a bit
ugly to remove all channels and put them back together
on every sync.