Settings: Add the channel logo visibility for the EPG
Some users might want to see the full channel name instead of
using the channel logo for easier distinction between channels.

An example of this would be the same channel using different
screen resolutions. In this case, they would use the same logo
altogether even though it is not the same channel "per-se".