Revamp Sign-in UI
- Use GuidedStepFragments in order to show a more consistent
UX when using the Leanback Support Library. They are
ordered between the Welcome, Email, Password, Loading and
Connected phases.
- Remove the old Auth Fragments since it wasn't all that pretty
for TV's.
- Replace the responsibility of AuthActivityTv to either know
whether executing itself or finishing itself. It should be up to
the demanding activities to call it if needed.
- AndroidManifest.xml: Change the default launching activity
to MainActivity.
- MainActivity: Determine whether the app is in TV UI mode
or not to show the correct fragments.
- AuthPresenterImpl: Set the default stream type to HLS.
- Create a Utilities class in order to simply code reuse
throughout classes.