AuthPresenter: Do MVP the right way (I hope)
In order to do the MVP pattern the right way, there shouldn't be
any dependency about Android (com.android.*) in the
presenters.

Because of this, there's a refactor that needs to be done so
here it is:

- Separate AsyncValidateInfo into an external class named
AsyncAuthValidateInfo
- In order to remove dependencies with SharedPreferences,
create an interface alongside it's implementation so it can be
passed through AsyncAuthValidateInfo
- AuthView: Replace onConnectionFailed() with
onWrongCredentialsReceived() since it is the only reason of
a failed connection so far as well as adding a new method
getPlayListLink().
- RichFeedUtil: Reduce Context dependency by overloading
getInputStream() so that it can only receive a url.

Some stuff could be made better, (getting the
playlist url in the view for example) but for
the simplicity/possible laziness, it has been
put there.