AceTvInputService: Adjust some errors
It might happen that some errors occur during a streaming. Some
of them have been recognized like BehindLiveWindowException.

Handle two more error cause which might be related to error
codes 403 (for multi-connections) and 502 (for bad gateways)
as well as errors for a timeout.

Add Toasts to notify the user about the situation.