AsyncAuthValidateInfo: Add more granular exception handling
Let's not assume that every error will result from not valid
credentials. In that case, let's verify each timeout. So far two
have been found which are:

- FileNotFoundException: Invalid credentials/possible issue with
A.C.E servers.
- SocketTimeoutException: Self explanatory. So far it was only
possible to achieve it with a large number of request at a given
time. (With at least 500 tries)

Add a method to AuthView to reflect the issue.