Settings: Add a debug log section
Most of the Android TV devices cannot be easily debuggable due
to their nature of being connected to a given device at all times.

For this reason, add an option in which, users can retrieve a log
from the application and transfer it easily.

Add v13 support libraries for runtime permissions methods
for Fragments.

Note that it can only record the current process's logs but it's
not a problem since I don't really care about other apps. (Only
Live Channels but that's for another time)