Add channel grouping for Live Channels
Add some basic Live Channels groups based on a given
program's genre.

It is most likely the only way of grouping different channels since
they are hardcoded in the open source variant of the app itself
here:

https://android.googlesource.com/platform/packages/apps/TV/+/master/src/com/android/tv/data/GenreItems.java

At first only sports and 24/7 channels are grouped. Others
might be added in the future.