CardViewPresenter: Decouple usage with Picasso
Only done as a thought for future expansion of the app, let's
decouple CardViewPresenter from Picasso so any possible library
could use possibly load images into an ImageCardView.

For this reason, create a new interface CardViewImageProcessor
which will leave a possibility of using any wanted library to load
an image.