AceChannelUtil: Verify if the channel is a live one in the playlist
An issue not seen with Sports subscription but with Premium
ones. When retrieving the channels, a premium user will get
VOD content in the playlist itself. That creates unnecessary
channels and thus should only be in the VOD section of the app.

Verify if the M3U entry is a live channel first before adding it
to the list of channels.