AceChannelUtil: Remove redundant checks on createChannelList()
Initially we were checking if a given channel had the same id as
the one in XmlTvParser or if it was the only one in it's kind.

However checking if a given channel has EPG is redundant since
the program are verified at a later stage.

It also improves on the speed of the computing going to
O(N²) to O(N).