XmlTvParser: Don't overwrite a program list if already existent
In the case of some IPTV providers, multiple channels can be
given the same EPG id. However, the channels might be seen
differently in XmlTvParser.

When initially parsing channels, it adds the programs to a
temporary list and then removes all the existing ones from the
initial program list iterator. This causes all other channels
having the same id to overwrite the saved channel list with an
empty one since there aren't any programs left in the given
iterator.

Fix this by only adding a given list if the id wasn't already
existent in mProgramMap.