AceJobService: Retrieve program list easily from XmlTvParser
The method of retrieving the programs at first was to check all
the programs and see if the given channel id from it matched
the EPG id for a channel. This was mostly due to getPrograms()
wasn't working perfectly in the first place.

Adapt AceJobService from the recent changes to XmlTvParser so
we can use it's getPrograms() method accordingly.

It should also offer some performance improvements as well.
(O(1) vs O(N))