ChannelLogo/UserInfo: Verify if the Activity is still up
Some users might quit the Activity before the AsyncTask has ran
the wanted part. If that's the case, the Activity from getActivity()
is null and return an exception.

Be sure that the Activity present by verifying if the Fragment has
been correctly added to it.

Some devices might be underpowered and therefore can't do all the
operations as fast. (For example the Amazon Fire TV Stick)