Implement Realm for on-device caching
In some cases, (Tv Catchup on a premium account) can take a
while to load (more than 10 seconds) and that might play a role
in the user experience.

Cache AvContents in a database powered by Realm for faster
retrieval of data.

Update the database if the values from a playlist is different.