Implement an in-app updater solution
Since we can't really use the Play Store for updating the app, let's
use QuantaWorks in order to replicate the functionality.

- Add AppUpdateGuidedFragment which will notify the user
concerning future updates of the application.
- Add AppUpdateDownloadGuidedFragment which will
download the file and install it if necessary.
- Add the required Views/Presenters/Implementations as made
for MVP
- Add AppUpdateGuidedFragment as a subsection in
AboutGuidedFragment
AndroidManifest.xml: Add Read storage permissions as well as
a definition for FileProvider (Required starting from Nougat)
- Constants: Add QuantaWorks URL's for Ace TV
- Add related string.xml entries