Settings: Use better listeners when updating channels
Using a toast to notify the user that given channels were syncing
wasn't the best option as he/she could simply go back to the
Live TV application without knowing if the changes were applied
correctly.

In that situation, use a custom GuidedStepFragment coupled
with a BroadcastReceiver to exit the required the activity and
leave a message for the user to restart the Live TV application
if not done.