Settings: Add a channel region filtering section
A request that have been asked is to be able to only watch live
TV from a specific region. Because Ace have a lot of channels,
(More than 900 for the premium package as of August 2017)
some might not want to go to all the list of channels.

Grouping them helped a bit in that regard but let's go a little
further than that.

Let's only have 3 regions at the moment: UK, USA/CA and
International. Some others can be added in the future as well.