Initial monorepo merge commit for Tsutaeru/xipl
Using a multi-repo approach for Tsutaeru and xipl can be useful
if one of the repos is a library itself. However, it was beginning to
be a bit harder to manage two repos at the same time. Moreover, I think
the scope of xipl is perhaps only restricted to Tsutaeru and thus could
be only used as an external module instead of a full fledged library.

Using a mono repo might be harder for viewing git's history since
it is not linear but the workflow will be improved. Also the old repos
should be available for archiving purposes just in case.