Change aipl/axipl to xipl
It is more representative to change everything to xipl since the library only
concerned XtreamCodes providers at the moment.

Also "Android Xtream Codes IPTV Provider Library" was a bit long so only
use "Xtream Codes IPTV Provider Library" as such. It might be changed in
the future if there is an issue concerning this.