Refactoring to xipl: Part 1
As a part of a newly created library for Xtream Codes based providers,
(xipl) let's adapt Ace TV to some of xipl's methods/classes

- settings.gradle: Compile xipl/tiflibrary from their given directory since
it will be easier to implement and update between it's revisions.
- Replace AceTvInputService by ProviderInputTvService
- Remove AcePlayer
- playback: Remove everything except PlaybackActivity and
PlaybackFragment and rename them accordingly for Ace TV.
- playback: Add AceVodPlaybackProperties for custom playback
properties for Ace.
- settings: Recreate the base fragments so they can be extended from
xipl's ProviderSettingsObjectAdapter and ProviderSettingsTvFragment.
- Rename acetvinputservice.xml to ace_tv_input_service.xml.