VodPlaybackFragment: Add support for seeking MPEG2-TS content
Due to MPEG-TS's implementation, there is no real way of seeking through
it with conventional means.

Let's use Abdallah Moussawi's implementation of seeking which is based
on the PCR value of the stream itself.

Since recorded contents on Xtream Codes generally gives the duration in
minutes, let's use it to our advantage for generating the required Extractor.

Also move ExoPlayerAdapter from the ui.vod package to the player one.