AsyncDownloadChangelog: Only verify against the latest version available
Previously, the loop would have continued until we found the last version
higher than the one installed which would be the absolute next version and
thus, probably not the most recent one.

Only retrieve the latest version and verify against it since we don't have a
need to verify versions.