VodTvSectionFragment: Revamp content updating
The earlier method of updating content was rather clumsy. The updated
content would only appear once a user would change sections. It was
a bit of gymnastics to get differences between content at the time.

With DiffUtil and DiffCallback in the Support Library, let's use
these new tools to update accordingly contents in a given section.