AsyncProcessAvContent: Adjust logic when fetching new content
In some cases, the M3U playlist might be empty when fetching new
content. (Caused by something on the external API)  This either
deletes the cache's contents or leave a section loading indefinitely.

If the cache does exist, don't touch anything in it and show an error
message if it doesn't.