auth: Adjust user login process
Since the app is aimed to TV users, the experience should be as simple
as possible. Using a long URL might be a bother to tap especially on
XtreamCodes based ones. Because of this, let's separate the process into:

- Base URL
- Username
- Password

This way, The user might not have that much to enter and it is less
error prone as well.