tiflibrary: XmlTvParser: Pass the correct Epg Id value
One could think that since id is a int, it could be represented the same
as a long. However, when autoboxed, the object is not the same at all
and it reflects in a Map.

In these cases, cast the id as a long in order to retrieve the list
of programs for a channel.