Use WeakReference for context in AsyncTask classes
Since a memory leak can occur when transferring contexts to AsyncTask
based classes, use a WeakReference that can permit a context to be
picked up by the GC. Ensure that all operations can only take place
when the context is not null.

ChannelLogoGuidedFragment: Don't sync the channels all over again if
we're removing them.