MainActivity: Recreate the Activity after auth is successful
On API 27, the mix of creating the layout alongside using Support Libraries
seems to create a state loss since we're executing the transaction after
the activity has it's instance state saved.

In that case, recreate the Activity so it can correctly configure the
required layout for MainTvFragment.

More details regarding the issue can be found here:
https://www.androiddesignpatterns.com/2013/08/fragment-transaction-commit-state-loss.html