ChannelLogoGuidedFragment: Ensure deletion with an active Fragment
When exiting the Fragment, the WeakReference used for the Context is
getting removed by the GC and this cause some channel logos not being
removed. In that case, let's show the user with a loading Fragment
and then close the Activity after the operation is done.

Since we already had EpgSyncLoadingGuidedFragment, let's use it as a
loading Fragment.