Remove region based filtering
Since there is a lot of providers offering their own set of groups,
providing regions might be in itself a arduous task to do.

In this case, simply remove the feature since it nearly won't be as
polished as giving the user the choice of managing his/her own groups.