xipl: Start Kotlin refactor
Since Kotlin's official support for Android, there has been a massive
uphold from the community. That was in part why Tsutaeru was rewritten
a little bit ago in Kotlin.

Rewriting the whole module from one language to another might be a bit
too much but in order to have more confidence in the code that is
written, it is a good idea to have unit tests that can verify certain
behaviors.

Let's use this refactor to then apply some testing practices. Begin it
with AvContentUtil which has the least dependencies to the Android
framework components.